move applyTransform
